pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-debugger-env"
version = "1.0.0"
description = "An OpenEnv RL environment for debugging Python code snippets"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "openenv-core",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.hatch.build.targets.wheel]
packages = ["code_debugger_env"]

[tool.uv.scripts]
server = "uvicorn server.app:app --host 0.0.0.0 --port 8000 --reload"
